+++ src/deposit.h
@@ -1,7 +1,7 @@
 #ifndef DEPOSIT_H
 #define DEPOSIT_H
 
-float sum_analiz (int days, float sum);
-int correct_analiz (int days, float sum);
+int sum_analiz (int days, int sum);
+int correct_analiz (int days, int sum);
 
 #endif
